Re-organized components folder- created a homepage folder to store files realted to the homepage there, created a SignUpPage folder and a Reusable components folder too. Modified certain path files
